+++ lib/ui/widget/alice_call_response_widget.dart
@@ -1,8 +1,8 @@
 import 'package:alice/model/alice_http_call.dart';
-import 'package:alice/utils/alice_constants.dart';
 import 'package:alice/ui/widget/alice_base_call_details_widget.dart';
-import 'package:better_player/better_player.dart';
+import 'package:alice/utils/alice_constants.dart';
 import 'package:flutter/material.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class AliceCallResponseWidget extends StatefulWidget {
   final AliceHttpCall call;
@@ -24,7 +24,6 @@ class _AliceCallResponseWidgetState
   static const _textContentType = "text";
 
   static const _kLargeOutputSize = 100000;
-  BetterPlayerController? _betterPlayerController;
   bool _showLargeBody = false;
   bool _showUnsupportedBody = false;
 
@@ -53,12 +52,6 @@ class _AliceCallResponseWidgetState
         ),
       );
     }
-  }
-
-  @override
-  void dispose() {
-    _betterPlayerController?.dispose();
-    super.dispose();
   }
 
   List<Widget> _buildGeneralDataRows() {
@@ -194,14 +187,6 @@ class _AliceCallResponseWidgetState
   }
 
   List<Widget> _buildVideoBodyRows() {
-    _betterPlayerController = BetterPlayerController(
-      const BetterPlayerConfiguration(aspectRatio: 16 / 9, fit: BoxFit.cover),
-      betterPlayerDataSource: BetterPlayerDataSource(
-        BetterPlayerDataSourceType.network,
-        _call.uri,
-      ),
-    );
-
     final List<Widget> rows = [];
     rows.add(
       Row(
@@ -214,9 +199,12 @@ class _AliceCallResponseWidgetState
       ),
     );
     rows.add(const SizedBox(height: 8));
-    rows.add(
-      BetterPlayer(controller: _betterPlayerController!),
-    );
+    rows.add(TextButton(
+      child: Text("Open video in web browser"),
+      onPressed: () async {
+        await launchUrl(Uri.parse(_call.uri));
+      },
+    ));
     rows.add(const SizedBox(height: 8));
     return rows;
   }